final optimized recommender system

+++ model/dataset.py
@@ -3,8 +3,8 @@ from torch.utils.data import Dataset
 
 class InteractionDataset(Dataset):
     def __init__(self, df):
-        self.users = torch.tensor(df['user'].values)
-        self.items = torch.tensor(df['item'].values)
+        self.users = torch.tensor(df['user'].values, dtype=torch.long)
+        self.items = torch.tensor(df['item'].values, dtype=torch.long)
         self.labels = torch.tensor(df['label'].values, dtype=torch.float32)
 
     def __len__(self):
